Optimizes data fetching with useCallback
Improves performance by using `useCallback` to memoize the `fetchBusiness` and `fetchFeedbacks` functions in the business details page and `fetchBusiness` and `generateQRCode` functions in the QR code generator page. This prevents unnecessary re-renders and re-fetches when the component updates, as the functions are only recreated when their dependencies change. Also, includes the business object as a dependency for generating QR code.